+++ server.go
@@ -606,7 +606,7 @@ func expirebackup() {
 }
 
 func printstats(name string, stat *syscall.Statfs_t) {
-	fmt.Printf("%-10s\t%s\t%s\t%s\t%.0f%%\t%s\n", Fstype(uint64(stat.Type)), Bytes(uint64(stat.Bsize)*stat.Blocks), Bytes(uint64(stat.Bsize)*(stat.Blocks-stat.Bavail)), Bytes(uint64(stat.Bsize)*stat.Bavail), 100-100*float64(stat.Bavail)/float64(stat.Blocks), name)
+	fmt.Printf("%-10s\t%s\t%s\t%s\t%.0f%%\t%s\n", Fstype(uint64(stat.Type)), Bytes(uint64(stat.Bsize)*stat.Blocks), Bytes(uint64(stat.Bsize)*(stat.Blocks-stat.Bavail)), Bytes(uint64(stat.Bsize)*stat.Bavail), 100-100*float32(stat.Bavail)/float32(stat.Blocks), name)
 }
 
 func df() {
@@ -640,7 +640,7 @@ func df() {
 		fmt.Println("Retention schedules:", schedules)
 		fmt.Println("Backup sets:", backups)
 		fmt.Printf("Data in vault: %s (%d files)\n", Bytes(uint64(size)), files)
-		fmt.Printf("Compression factor: %.1f\n", float64(size)/(float64(vstat.Bsize)*float64(vstat.Blocks-vstat.Bavail)))
+		fmt.Printf("Compression factor: %.1f\n", float32(size)/(float32(vstat.Bsize)*float32(vstat.Blocks-vstat.Bavail)))
 	}
 }
